添加：添加角色管理等功能

+++ dao/dao-api/src/main/java/org/hengsir/icma/dao/LeftMenuWriteDao.java
@@ -29,7 +29,7 @@
        * @param id 菜单id
        * @return true： 修改成功,false： 修改失败
        */
-      public boolean delete(Long id);
+      public boolean delete(int id);
 
       /**
        * 修改菜单的权限。
